Add execute and shutdown commands
As well as sending a file with the send command (previously sendfile),
the user can now request the server to execute a file.  The server can
receive any number of commands and will shut down if the client closes
its socket.  In this case, it will listen for further commands until it
receives a shutdown command.