Add support for "execute" and "send" argument keywords
Now that the /K switch is implemented in RArgs, update the handling of
the execute and send commands to use it.  This enables the user to
explicitly specify the name of the file to be executed or sent, rather
than being hard coded for execute, and extracted from the default
argument list for send, which were both always temporary hacks.