Implement support for capturing output
When commands are remotely executed, their stdin, stdout and stderr are
now redirected so that the output can be captured and sent back to the
client computer.  This initial implementation provides support only for
Windows server builds.  All clients builds, however, will display the
captured output from the execution of commands on a Windows server.