Make RSocket read or write all bytes requested
It is now guaranteed, when calling RSocket::read() or RSocket::write(),
that the entire amount of bytes requested will be transferred.  The
socket class will now loop around if a call to recv() or send() on the
underlying BSD socket returns less than the requested amount, even if
this results in blocking.

Futhermore, RSocket::read() accepts a flag to disable this behaviour and
to go back to the original partial read behaviour if required.