Use exception handling for sockets
RSocket now throws exceptions if read() or write() fails due to not
being able to transfer the data requested or the socket being closed by
the remote host.  This reduces the granularity of the error messages
that are displayed when an error occurs but also shortens the code.  The
now missing errors would only be displayed if the stream had somehow
become out of sync, which is unlikely and unrecoverable, so it is not
much of a loss of functionality.