Add support for new commands
New commands required for the implementation of remote file access in
Brunel have been added.  The delete command allows the deletion of a
remote file, the rename command allows the renaming of a remote file
and the fileinfo command allows the user to query the remote server
for detailed information (such as size and type) about a specific file
in the remote filesystem.