Split listen() and accept() calls
Until now, RADRunner clients would make a single connection to the
server and would then exit.  With the reimplementation of the protocol
for Brunel, connections can now be made very quickly, one after another,
which requires that the same socket be used for multiple connections.
Thus, listen() is now only called once, and accept() is called for each
connection from a client, to avoid race conditions when connections come
from the client very quickly.